frame: change usec type from u64 to u32 in quic_usec_to_msec()
This patch is to fix the compiling error on i686:

  ERROR: modpost: "__divdi3" [net/quic/quic.ko] undefined!
  ERROR: modpost: "__udivdi3" [net/quic/quic.ko] undefined!
  make[2]: *** [scripts/Makefile.modpost:147: Module.symvers] Error 1
  make[1]: *** [/home/net-next/Makefile:2051: modpost] Error 2
  make: *** [Makefile:248: __sub-make] Error 2

+++ modules/net/quic/frame.c
@@ -2430,7 +2430,7 @@ static u8 *quic_frame_put_address(u8 *p, u16 id, union quic_addr *addr,
 	return p;
 }
 
-static inline u64 quic_usec_to_msec(u64 usec)
+static inline u64 quic_usec_to_msec(u32 usec)
 {
 	return DIV_ROUND_UP(usec, 1000);
 }
